% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASF.R
\name{ASF}
\alias{ASF}
\title{Apply Alternate-Sequential Filter}
\usage{
ASF(X, A)
}
\arguments{
\item{X}{A single input image or a list of input images}

\item{A}{List of structuring elements}
}
\value{
Images obtained after applying the ASF with structuring elements A
}
\description{
Apply an Alternate-Sequential Filter
}
\examples{
y <- ASF(input_digits[[1]],list(rbind(c(1,0,0),c(0,1,0),c(0,0,1)),
         rbind(c(0,1,0),c(1,1,1),c(0,1,0))))
image_bw(y)
}
