% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asf_sg16_sg16.R
\docType{data}
\name{asf_sg16_sg16}
\alias{asf_sg16_sg16}
\title{ASF-SG16-SG16 DMNN architecture}
\format{
An object of class \code{list} of length 4.
}
\usage{
asf_sg16_sg16
}
\description{
The Morphological Computational Graph of a Discrete Morphological Neural Network with a ASF layer with width 3, a sup-generating
 layer with width 16 and window 3x3, and a sup-generating layer with width 16 and window 3x3
}
\details{
A list of:
* Test Test
\itemize{
\item{edges}{ Matrix with edges of the Morphological Computational Graph}
\item{vertices}{ A vector with the vertices of the Morphological Computational Graph}
\item{compute}{ A vector with the computation type of each edge of the Morphological Computational Graph}
\item{param}{ A list with parameters of a realization of the Discrete Morphological Neural Network Parameters}
}
}
\keyword{datasets}
