% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_DMNN.R
\name{check_DMNN}
\alias{check_DMNN}
\title{Check a DMNN architecture}
\usage{
check_DMNN(vertices, edges, compute, param)
}
\arguments{
\item{vertices}{A vector of vertices}

\item{edges}{A matrix of edges}

\item{compute}{A vector with the computation type of the vertices}

\item{param}{List with the parameters of a realization of the DMNN}
}
\value{
Logical indicating if the DMNN represented by the given Morphological Computational graph is valid
}
\description{
Check if a Discrete Morphological Neural Network architecture is valid
}
\examples{
check_DMNN(asf_sg16_sg16$vertices,asf_sg16_sg16$edges,asf_sg16_sg16$compute,asf_sg16_sg16$param)
}
