% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMNN.R
\name{DMNN}
\alias{DMNN}
\title{Apply a DMNN}
\usage{
DMNN(X, edges, compute, param)
}
\arguments{
\item{X}{A single input image}

\item{edges}{A matrix with the edges}

\item{compute}{A vector with the computation type of vertices}

\item{param}{List with the parameters of vertices computation}
}
\value{
Images obtained after applying the given DMNN to X
}
\description{
Apply a Discrete Morphological Neural Network
}
\examples{
y <- DMNN(input_digits[[1]],asf_sg16_sg16$edges,asf_sg16_sg16$compute,asf_sg16_sg16$param)
image_bw(y)
}
