% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_asf.R
\name{error_asf}
\alias{error_asf}
\title{Error of Alternate-Sequential Filter}
\usage{
error_asf(X, Y, A, type = "mae")
}
\arguments{
\item{X}{List of input images}

\item{Y}{List of output images}

\item{A}{List of structuring elements}

\item{type}{Error type. Should be "iou" or "mae"}
}
\value{
Error of ASF with structuring elements A on sample (X,Y)
}
\description{
Calculate the error of an Alternate-Sequential Filter
}
\examples{
error_asf(X = input_digits[[1]],Y = output_digits[[1]],
          A = list(rbind(c(1,0,0),c(0,1,0),c(0,0,1)),rbind(c(0,1,0),c(1,1,1),c(0,1,0))))
}
