% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_DMNN.R
\name{error_DMNN}
\alias{error_DMNN}
\title{Error of a DMNN}
\usage{
error_DMNN(X, Y, edges, compute, param, type = "mae")
}
\arguments{
\item{X}{List of input images}

\item{Y}{List of output images}

\item{edges}{A matrix with the edges}

\item{compute}{A vector with the computation type of vertices}

\item{param}{List with the parameters of vertices computation}

\item{type}{Error type. Should be "iou" or "mae"}
}
\value{
Error of the DMNN on sample (X,Y)
}
\description{
Calculate the error of a Discrete Morphological Neural Network
}
\examples{
error_DMNN(X = input_digits,Y = output_digits,edges = asf_sg16_sg16$edges,
           compute = asf_sg16_sg16$compute,
           param = asf_sg16_sg16$param)
}
