% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_supgen.R
\name{error_supgen}
\alias{error_supgen}
\title{Error of sup-generating decomposed operator}
\usage{
error_supgen(X, Y, A, B, type = "mae")
}
\arguments{
\item{X}{List of input images}

\item{Y}{List of output images}

\item{A}{List of lower limits of intervals}

\item{B}{List of upper limits of intervals}

\item{type}{Error type. Should be "iou" or "mae"}
}
\value{
Error of operator with kernel given by the union of the intervals represented by [A,B]
}
\description{
Calculate the error of an operator represented by a sup-generating decomposition
}
\examples{
error_supgen(X = input_digits[[1]],Y = output_digits[[1]],A = rbind(c(0,1,0),c(1,1,1),c(0,1,0)),
             B = rbind(c(1,1,1),c(1,1,1),c(1,1,1)))
}
