% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_bw.R
\name{image_bw}
\alias{image_bw}
\title{Plot a binary image}
\usage{
image_bw(X)
}
\arguments{
\item{X}{A binary matrix}
}
\value{
A ggplot object with a black and white image
}
\description{
Plot a binary image from a matrix
}
\examples{
image_bw(input_digits[[1]])
}
