% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_arch_DMNN.R
\name{initialize_arch_DMNN}
\alias{initialize_arch_DMNN}
\title{Initialize DMNN architecture}
\usage{
initialize_arch_DMNN(compute, width, density = 0.3, random = "random")
}
\arguments{
\item{compute}{A vector with the computation type of vertices}

\item{width}{Width of each computation}

\item{density}{Expected density of ones in the random parameters}

\item{random}{Type of initialization. Should be "random", "random_identity" or "identity"}
}
\value{
List with initial parameters for the DMNN architecture
}
\description{
Randomly initialize the parameters of a Discrete Morphological Neural Network architecture
}
\examples{
param <- initialize_arch_DMNN(asf_sg16_sg16$compute,rep(3,length(asf_sg16_sg16$compute)))
}
