% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_digits.R
\docType{data}
\name{input_digits}
\alias{input_digits}
\title{Noised black and white digits}
\format{
An object of class \code{list} of length 10.
}
\usage{
input_digits
}
\description{
A list of black and white digits with salt and pepper noise
}
\details{
A list with ten 56 x 56 binary matrices
}
\keyword{datasets}
