% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDGA.R
\name{LGDA}
\alias{LGDA}
\title{Lattice Gradient Descent Algorithm for DMNN}
\usage{
LGDA(
  X,
  Y,
  Xval = NULL,
  Yval = NULL,
  b = 1,
  epochs = 10,
  vertices,
  compute,
  edges,
  param,
  type = "mae",
  sample = F,
  cores = 1
)
}
\arguments{
\item{X}{List of input training images}

\item{Y}{List of output training images}

\item{Xval}{List of input validation images}

\item{Yval}{List of output validation images}

\item{b}{Batch size}

\item{epochs}{Number of training epochs}

\item{vertices}{A vector of vertices}

\item{compute}{A vector with the computation type of vertices}

\item{edges}{A matrix with the edges}

\item{param}{List with initial parameters of vertices computation}

\item{type}{Error type. Should be "iou" or "mae"}

\item{sample}{Number of neighbors to sample or FALSE if it should not sample}

\item{cores}{Number of cores for parallel computing.}
}
\value{
param_global Parameters of the DMNN with the least training error obtained during the training

last_param Parameters of the last DMNN

trace A data frame with the trace of the algorithm
}
\description{
Lattice Gradient Descent Algorithm for training general Discrete Morphological Neural Networks
}
\examples{
train <- LGDA(X = list(input_digits[[1]]),Y = list(output_digits[[1]]),Xval = input_digits,
              Yval = output_digits,
              b = 10,epochs = 1,vertices = asf_sg16_sg16$vertices,compute = asf_sg16_sg16$compute,
              edges = asf_sg16_sg16$edges,param = asf_sg16_sg16$param,sample = 50)
}
\references{
D. Marcondes, J. Barrera. Discrete Morphological Neural Networks. 2023.
}
