% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_digits.R
\docType{data}
\name{output_digits}
\alias{output_digits}
\title{Boundary of black and white digits}
\format{
An object of class \code{list} of length 10.
}
\usage{
output_digits
}
\description{
Boundary of the black and white digits in the input_digits list
}
\details{
A list with ten 56 x 56 binary matrices
}
\keyword{datasets}
