% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_matrix.R
\name{pad_matrix}
\alias{pad_matrix}
\title{Pad a matrix with zeros}
\usage{
pad_matrix(m, s)
}
\arguments{
\item{m}{A matrix}

\item{s}{Padding size}
}
\value{
Matrix padded with s zeros
}
\description{
Pad a matrix with zeros on the left, right, above and below
}
\examples{
pad_matrix(matrix(1,3,3),1)
}
