% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_architecture.R
\name{plot_architecture}
\alias{plot_architecture}
\title{Plot a Morphological Computational Graph}
\usage{
plot_architecture(vertices, compute, edges)
}
\arguments{
\item{vertices}{A vector of vertices}

\item{compute}{A vector with the computation type of vertices}

\item{edges}{A matrix with the edges}
}
\value{
A plot with the architecture graph
}
\description{
Plot the Morphological Computational Graph of a Discrete Morphological Neural Network
}
\examples{
plot_architecture(vertices = asf_sg16_sg16$vertices,compute = asf_sg16_sg16$compute,
                 edges = asf_sg16_sg16$edges)
}
