% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_generating.R
\name{sup_generating}
\alias{sup_generating}
\title{Apply a sup-generating operator}
\usage{
sup_generating(X, A, B)
}
\arguments{
\item{X}{A single input image or a list of input images}

\item{A}{Lower limit of interval}

\item{B}{Upper limit of interval}
}
\value{
Images obtained after applying the sup-generating operator with interval [A,B].
}
\description{
Apply a sup-generating operator to a single image or a list of images
}
\examples{
y <- sup_generating(X = input_digits[[1]],A = rbind(c(0,1,0),c(1,1,1),c(0,1,0)),
                    B = rbind(c(1,1,1),c(1,1,1),c(1,1,1)))
image_bw(y)
}
