% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woperator.R
\name{woperator}
\alias{woperator}
\title{Apply sup-generating decomposed operator}
\usage{
woperator(X, A, B)
}
\arguments{
\item{X}{A single input image or a list of input images}

\item{A}{List of lower limits of intervals}

\item{B}{List of upper limits of intervals}
}
\value{
Images obtained after applying the sup-generated operator with kernel given by the union of the intervals [A,B]
}
\description{
Apply the operator represented by a sup-generating decomposition
}
\examples{
A <- list(rbind(c(0,1,0),c(1,1,1),c(0,1,0)),rbind(c(1,0,0),c(0,1,0),c(0,0,1)))
B <- list(rbind(c(1,1,1),c(1,1,1),c(1,1,1)),rbind(c(1,1,1),c(1,1,1),c(1,1,1)))
y <- woperator(input_digits[[1]],A,B)
image_bw(y)
}
